Fix a crash when an unsafe lazy is accessed from multiple threads
Currently, the unsafe lazy can throw an exception under certain
conditions if accessed from multiple threads.

This is, of course, a valid behavior seeing as it's unsafe. But with a
slight modification, it can be changed to not throw.

The documentation has been kept the same to keep the implementation
details loose.