Use -f to avoid complains when script is run twice and file does not exist
The last step of the script deletes the verification-metadata.xml file.
This does not need to be recursive delete because it is a file not
directory. This should be -f because we do not want to complain if the
script is run twice and the file has already been deleted by the
previous run.

Change-Id: I867838a940fa6ef7320ee2df417134634160a0f4 ( https://android-review.googlesource.com/q/I867838a940fa6ef7320ee2df417134634160a0f4 )

Moved from: https://github.com/androidx/androidx/commit/10f503b047d151036e0a1bfe5d03519a9647789f