Fix stdlib annotation enum deserialization issue
From IntelliJ, when we access to an enum entry of a Kotlin stdlib
annotation from other libraries, `StubBasedAnnotationDeserializer`
generates `FirFromMissingDependenciesNamedReference`, which causes a
type resolution issue for FIR2IR. This issue was fixed by
a workaround commit 7db2fc522e9e333e69b92e46881214a040ac80c3 only for
the stand-alone compiler. This commit uses `FirBuiltinSymbolProvider`
for `LLStubBasedLibrarySymbolProviderFactory` to fix this issue for
IDEA as well.

^KT-61757