IR text tests: Verify IR signatures in Native KLIBs
This commit adds additional checks to the existing Native IR text tests.

1. Verify that the signatures that are computed for IR nodes during
compilation are exactly the same as the signatures that are actually
stored in KLIB. Do it both for K1 and K2 compilers.

2. Verify that the signatures that are computed for IR nodes during
compilation with K1 compiler are exactly the same as the signatures
that can be lately computed by LazyIr (K1) or Fir2LazyIr (K2) built
on top of KLIB's metadata.

3. Verify that the signatures that are computed for IR nodes during
compilation with K2 compiler are exactly the same as the signatures
that can be lately computed by Fir2LazyIr (K2) built
on top of KLIB's metadata.

^KT-58239