[K/N] Always add Objective-C category members to class in the same file
By default, Obj-C category members are imported as Kotlin extensions.
In many cases it is fine, however, it has the following drawbacks:
1. If a class member is moved to a category, it becomes a breaking
   change for Kotlin.
2. In some cases it should be possible to override methods that are
   declared in categories. See KT-56001.

There is a workaround: `objcClassesIncludingCategories` property in a
`.def` file makes `cinterop` expand listed classes with members of
categories defined in the same file as the class itself.

It is still far from perfect: it requires manual intervention and has
some usability problems apart from that.

This commit enables this behaviour for all Objective-C classes.
In other words, this commit makes `cinterop` tool expand any Objective-C
class with members of categories defined for this class in the same file
as the class itself.

^KT-56001 Fixed