[K/N] Deprecate Obj-C factory when including it also as a constructor
`cinterop` can copy Objective-C category methods to classes. In
particular, it can do that for `init` methods, making them proper class
constructors.
When copying regular methods and properties this way, `cinterop` marks
original generated extension declarations as deprecated. But it didn't
do that for `init` methods.

This commit fixes this inconsistency, marking the original extension
`create` function deprecated when copying it the class as a constructor.

^KT-65685 Fixed