K2: Fix internal error in FIR2IR caused by PCLA + String interpolation
There's an implicit contract in PCLA that the statement-level call
should be postponed iff it has something to be postponed inside.

And that contract didn't work well for string interpolation containing
some postponed calls.

Thus, we haven't run a completion results writing for them properly,
thus leaving type parameters (K from synthetic call) for expression
types instead of an inferred substituted type.

NB: In the original case, `x!!` is being postponed because for a different
reason we postpone all operator calls right now.

NB2: Basically, we might have postponed just any call, but that might
lead to performance issues because that would lead to huge constraint
systems, while in many cases a lot of calls are clearly trivial
and can't be connected to other calls in PCLA lambda.

^KT-65341 Fixed