Call startReplaceGroup instead of startReplaceableGroup
Changes the code generation to call `startReplaceGroup`
when it is availiable, instead of `startReplaceableGroup`.
This is a trigger to the runtime that this compiler generates
code that is correct if the replae groups are replaced
instead of treated as movable groups.

Prior versions of the compiler plugin did not generated the
correct code around loops that only contain composable
lambdas which caused the groups to become missaligned and
caused delete/inserts to be generated when a single delete
or insert would have been sufficient. Treating replace
groups as movable groups was resilient to the replace groups
being misaligned (at the cost of using a diff algorithm)
which hid this issue.

Bug: 322386428 ( https://issuetracker.google.com/issues/322386428 )
Test: ./gradlew :compose:c:c-h:i-t:tDUT

Change-Id: I346d9a67ed36a7350b40f05486f15d6956f85965 ( https://android-review.googlesource.com/q/I346d9a67ed36a7350b40f05486f15d6956f85965 )

Moved from: https://github.com/androidx/androidx/commit/5e3ef979c953b25e24241b334d61a951d6bcdf54