K2: Attach function parameter annotation to function expression
When a function parameter has an annotation, the exist FIR2IR does not
attach the annotation to the function expression (lambda argument),
which results in a missing annotation. It is critical to the Compose,
because the compose compiler / compose runtime rely on the Composable
annotation.

This commit attaches the annotation of a function parameter to the
function expression used for the argument (like lambda) to fix this
issue.

^KT-64994 Fixed