[K/N] Optimize searching for Obj-C categories declared in the same file
While indexing Objective-C headers, `cinterop` also looks for Obj-C
categories declared in the same files as their classes (see e.g.
KT-56001).

Currently, the implementation of this is inefficient -- it traverses the
entire translation unit for a single Objective-C class.

Luckily, this is enabled only for a few classes in total, so this
inefficiency doesn't affect the performance. But this is going to change
soon.

This commit optimizes the implementation by caching the translation
unit traversal results, thus requiring only a single traversal of the
involved translation units.

^KT-56001