[K/N] Make cinterop findObjCCategoriesInSameFilesAsClasses check TUs
findObjCCategoriesInSameFilesAsClasses iterates over translation units
containing encountered classes. The problem is that it is still a bit
fragile, because it is easy to have a disposed translation unit recorded
there.
As a way to make such problems easier to discover, this commit makes
findObjCCategoriesInSameFilesAsClasses check that all involved
translation units are still "valid" (= owned by a valid UnitsHolder).

^KT-56001