[Gradle] Report warning when there is multiple source roots in compilation
This issue should warn users about possible problems in K2 compiler.
Because in this setup there will be no symbols visibility between
those source set roots as it used to be in K1.
There is an assumption that this case usually appears in the code
generation setup. And users might not experience any problems before.

^KT-64913 Verification Pending