[IR] Drop IrIfThenElseImpl
in order to simplify IR classes for future changes.

IrIfThenElseImpl class was an additional implementation of IrWhen,
besides the main IrWhenImpl. The only difference from IrWhenImpl
was another implementation of a list, which likely has little
impact on performance. What's more, it is planned to replace all such
lists with a new, custom implementation, at which point those classes
would become identical.

#KT-65773 Related